Delivery in Auto working

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Helper/Chassis.java
@@ -130,7 +130,7 @@ public class Chassis {
         }
     }
 
-    public void Drive(double speed, int distance) {
+    public void Drive(double speed, float distance) {
 
         runtime.reset();
         timeout_ms = 10000;
